requirements.txt
fastapi
uvicorn
torch
transformers
natasha
python-Levenshtein
nltk
python-dotenv
gradio
openai
requests
